+++ CHANGELOG.md
@@ -4,4 +4,6 @@ All notable changes to the Zowe Client Python SDK will be documented in this fil
 
 ## Recent Changes
 
+- Feature: Added a CredentialManager class to securely retrieve values from credentials and manage multiple credential entries on Windows [#134](https://github.com/zowe/zowe-client-python-sdk/issues/134)
+- Feature: Added method to load profile properties from environment variables
 - Feature: Added method to Save profile properties to zowe.config.json file